Create Search folder, move components into folder.
